refactor(AFA-143): reorganize test files and update test configurations
